Adjustment

+++ app/controllers/likes_controller.rb
@@ -7,13 +7,10 @@ class LikesController < ApplicationController
     @post.create_notification_like!(current_user)
     redirect_to posts_path, success: t('defaults.message.liked', item: Like.model_name.human)
   end
-  
 
   def destroy
     like = Like.find_by(user_id: current_user.id, post_id: params[:id])
-    if like
-      like.destroy
-    end
+    like.destroy if like
     redirect_to posts_path, danger: t('defaults.message.not_liked', item: Like.model_name.human)
   end
